+++ app/definitions/field-validators/pre/multiple-job-select.js
@@ -1,0 +1,17 @@
+const {
+  validationRules: r,
+  simpleFieldValidation: sf,
+} = require('@dwp/govuk-casa');
+
+const fieldValidators = {
+  selectJob: sf([
+    r.required.make({
+      errorMsg: {
+        summary: 'multiple-job-select:errors.required',
+        inline: 'multiple-job-select:errors.required',
+      },
+    }),
+  ]),
+};
+
+module.exports = fieldValidators;