+++ app/routes/account/about-your-grant.js
@@ -1,5 +1,6 @@
 const {
-  ACCOUNT_CONTEXT_PATH,
+  GRANT_CONTEXT_PATH,
+  GRANT_ROOT_URL
 } = require('../../config/uri');
 const filterGrantsForActiveOnly = require('../../lib/utils/filter-claims');
 
@@ -13,10 +14,27 @@ const buildEmployersString = (employers) => {
 
 module.exports = (casaApp) => {
   const getPage = (req, res) => {
-    const grantAwards = req.casa.journeyContext.getDataForPage('__hidden_account__').account.elements;
-    const activeClaims = filterGrantsForActiveOnly(grantAwards);
+
+    const { job } = { ...req.session.selectedJob };
+    const { grants } = { ...req.session.grantsToShow };
+
+    let activeClaims;
+
+    if (grants === undefined) {
+      activeClaims = filterGrantsForActiveOnly(
+        req.casa.journeyContext.getDataForPage('__hidden_account__').account.elements);
+    } else {
+      if (grants.length == 1) {
+        activeClaims = filterGrantsForActiveOnly(grants);
+        res.locals.casa.journeyPreviousUrl = `${GRANT_ROOT_URL}/multiple-grant-select`;
+      } else {
+        activeClaims = filterGrantsForActiveOnly(grants.filter((grant) => grant.company == job));
+        res.locals.casa.journeyPreviousUrl = `${GRANT_ROOT_URL}/multiple-job-select`;
+      }
+    }
 
     res.locals.grantAwards = activeClaims;
+
 
     const employers = [...new Set(activeClaims.map((grant) => grant.company))];
 
@@ -24,7 +42,7 @@ module.exports = (casaApp) => {
     return res.render('pages/account/about-your-grant.njk');
   };
 
-  casaApp.router.get(`${ACCOUNT_CONTEXT_PATH}/grant-summary`, casaApp.csrfMiddleware, getPage);
+  casaApp.router.get(`${GRANT_CONTEXT_PATH}/grant-summary`, casaApp.csrfMiddleware, getPage);
 
   return {
     getPage,